requirements.txt
alabaster==0.7.10
alembic==0.9.3
aniso8601==1.2.1
appnope==0.1.0
Babel==2.4.0
bcrypt==3.1.3
bleach==2.0.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
click==6.7
coverage==4.4.1
decorator==4.1.2
defusedxml==0.5.0
Django==1.9.13
django-appconf==1.0.2
django-bootstrap-breadcrumbs==0.8.2
django-bower==5.2.0
django-compressor==2.1.1
django-cors-headers==2.1.0
django-filter==1.0.4
django-formtools==1.0
django-mobile==0.7.0
django-recaptcha==1.3.1
django-sortedm2m==1.4.0
django-tastypie==0.14.0
djangorestframework==3.2.5
docutils==0.13.1
easy-thumbnails==2.4.1
flake8==3.3.0
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.0
Flask-Migrate==2.0.4
flask-paginate==0.5.0
flask-restplus==0.10.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-Testing==0.6.2
gunicorn==19.7.1
html5lib==0.999999999
icalendar==3.11.5
idna==2.5
imagesize==0.7.1
invoke==0.19.0
ipdb==0.10.3
ipython==6.1.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
nose==1.3.7
oauthlib==2.0.2
olefile==0.44
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
psycopg2==2.7.1
ptyprocess==0.5.2
py-bcrypt==0.4
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.5.0
Pygments==2.2.0
PyJWT==1.5.2
python-dateutil==2.6.1
python-editor==1.0.3
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2017.2
rcssmin==1.0.6
reportlab==3.4.0
requests==2.18.1
requests-oauthlib==0.8.0
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
social-auth-app-django==1.2.0
social-auth-core==1.4.0
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
SQLAlchemy==1.1.11
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
traitlets==4.3.2
urllib3==1.21.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
wger==1.8
